Revert "SystemSan: arbitrary DNS resolution detection (#8448)"
This reverts commit 98eda2bc0d3d43ebd4addbe901972c0d3b5f4cb9.
