+++ infra/experimental/SystemSan/Makefile
@@ -2,9 +2,9 @@
 CXX     = clang++
 CFLAGS = -std=c++17 -Wall -Wextra -O3 -g3
 
-all: clean SystemSan target target_file target_dns
+all: clean SystemSan target target_file
 
-SystemSan: SystemSan.cpp inspect_dns.cpp inspect_utils.cpp
+SystemSan: SystemSan.cpp
 	$(CXX) $(CFLAGS) -lpthread -o $@ $^
 
 target: target.cpp
@@ -13,13 +13,9 @@ target: target.cpp
 target_file: target_file.cpp
 	$(CXX) $(CFLAGS) -fsanitize=address,fuzzer -o $@ $^
 
-target_dns: target_dns.cpp
-	$(CXX) $(CFLAGS) -fsanitize=address,fuzzer -o $@ $^
-
 test:  all vuln.dict
 	./SystemSan ./target -dict=vuln.dict
 	./SystemSan ./target_file -dict=vuln.dict
-	./SystemSan ./target_dns -dict=vuln.dict
 
 pytorch-lightning-1.5.10:
 	cp SystemSan.cpp PoEs/pytorch-lightning-1.5.10/; \
@@ -34,4 +30,4 @@ node-shell-quote-v1.7.3:
 	docker run -t systemsan_node-shell-quote:latest;
 
 clean:
-	rm -f SystemSan /tmp/tripwire target target_file target_dns
+	rm -f SystemSan /tmp/tripwire target target_file
