--- infra/experimental/SystemSan/inspect_dns.h
@@ -1,26 +1,0 @@
-/*
- * Copyright 2022 Google LLC
-
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
-
- *      http://www.apache.org/licenses/LICENSE-2.0
-
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-/* A detector that uses ptrace to identify DNS arbitrary resolutions. */
-
-
-/* POSIX */
-#include <unistd.h>
-
-/* Linux */
-#include <sys/ptrace.h>
-
-
-void inspect_dns_syscalls(pid_t pid, const user_regs_struct &regs);
