+++ infra/experimental/SystemSan/vuln.dict
@@ -1,3 +1,2 @@
 "/tmp/tripwire"
 "/fz/"
-"f.z"
